requirements.txt
Django==1.11.26
django-autoslug==1.9.3
django-formtools==2.0
django-crispy-forms==1.6.1
django-extensions==1.8.0
django-file-form==0.2.3
django-guardian==1.4.8
django-model-utils==3.0.0
django-sitetree==1.8.0
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
django-downloadview==1.9
django-debug-toolbar==1.8
django-debug-panel==0.8.3
django-mptt==0.8.7
django-ordered-model==1.4.1
django-ajax-selects==1.6.1
django-filter
pathlib
shapely
GeoAlchemy2
lxml
psycopg2
psycopg2-binary
pillow
owslib
defusedxml
SQLAlchemy
django-import-export
coverage
urllib3
python-magic
celery==3.1.15
django-celery

# package depends on libgdal-dev package => version have to match, e.g. isntall only 
# the version of the installed dev version. Need also to follow env var setup before install as described in:
# https://gis.stackexchange.com/questions/28966/python-gdal-package-missing-header-file-when-installing-via-pip
# to allow finding of libs and includes.
GDAL
